use latest swift-rs
