+++ apps/desktop/src-tauri/Cargo.toml
@@ -19,13 +19,13 @@ tracing = "0.1.36"
 serde = "1.0.145"
 
 [target.'cfg(target_os = "macos")'.dependencies]
-swift-rs = { git = "https://github.com/Brendonovich/swift-rs.git", branch = "autorelease" }
+swift-rs.workspace = true
 
 [build-dependencies]
 tauri-build = { version = "1.1.1", features = [] }
 
 [target.'cfg(target_os = "macos")'.build-dependencies]
-swift-rs = { git = "https://github.com/Brendonovich/swift-rs.git", branch = "autorelease", features = [
+swift-rs = { workspace = true, features = [
   "build",
 ] }
