+++ apps/desktop/src-tauri/src/macos/webview.rs
@@ -1,4 +1,5 @@
 use super::native::NSObject;
-use swift_rs::*;
 
-pub_swift_fn!(reload_webview(webview: NSObject));
+extern "C" {
+	pub fn reload_webview(webview: &NSObject);
+}
