+++ apps/desktop/src-tauri/src/main.rs
@@ -80,8 +80,9 @@ async fn main() -> Result<(), Box<dyn Error>> {
 					use macos::*;
 
 					let window = window.ns_window().unwrap();
-					set_titlebar_style(window, true, true);
-					blur_window_background(window);
+
+					unsafe { set_titlebar_style(&window, true, true) };
+					unsafe { blur_window_background(&window) };
 				}
 			});
